Refactor AssetTransferApi integration tests
