+++ src/testHelpers/mockWeightInfo.ts
@@ -1,0 +1,10 @@
+// Copyright 2023 Parity Technologies (UK) Ltd.
+
+export const mockWeightInfo = {
+	weight: {
+		refTime: '133179000',
+		proofSize: '0',
+	},
+	class: 'Normal',
+	partialFee: '171607466',
+};
